Parse doubles correctly, instead of as ints.

+++ hubitat2prom/HubitatModels/HubitatDeviceAttributes.cs
@@ -5,7 +5,7 @@ namespace hubitat2prom.HubitatModels;
 public class HubitatDeviceAttributes
 {
     public string name { get; set; }
-    public OneOf<int, string>? currentValue { get; set; }
+    public OneOf<double, string>? currentValue { get; set; }
     public string dataType { get; set; }
     public string[] values { get; set; }
 }
